задание 5.модуляция(версия 1)
